requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-braces==1.11.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-extensions==1.7.6
django-waffle==0.11.1
django==1.10.5
djangorestframework-jwt==1.9.0
djangorestframework==3.5.4
gunicorn==19.6.0
psycopg2==2.6.2
PyJWT==1.4.2              # via djangorestframework-jwt
six==1.10.0               # via django-environ, django-extensions
sqlparse==0.2.2           # via django-debug-toolbar
